+++ service/cleanroomsml/types/types_exported_test.go
@@ -60,3 +60,21 @@ func ExamplePrivacyBudgets_outputUsage() {
 }
 
 var _ []types.AccessBudget
+
+func ExampleWorkerComputeConfigurationProperties_outputUsage() {
+	var union types.WorkerComputeConfigurationProperties
+	// type switches can be used to check the union value
+	switch v := union.(type) {
+	case *types.WorkerComputeConfigurationPropertiesMemberSpark:
+		_ = v.Value // Value is map[string]string
+
+	case *types.UnknownUnionMember:
+		fmt.Println("unknown tag:", v.Tag)
+
+	default:
+		fmt.Println("union is nil or unknown type")
+
+	}
+}
+
+var _ map[string]string