+++ service/storagegateway/go_module_metadata.go
@@ -3,4 +3,4 @@
 package storagegateway
 
 // goModuleVersion is the tagged release for this module
-const goModuleVersion = "1.43.9"
+const goModuleVersion = "1.43.10"