+++ service/bcmdashboards/CHANGELOG.md
@@ -1,3 +1,7 @@
+# v1.0.19 (2026-01-09)
+
+* **Dependency Update**: Updated to the latest SDK module versions
+
 # v1.0.18 (2025-12-08)
 
 * **Dependency Update**: Updated to the latest SDK module versions