+++ service/sagemaker/CHANGELOG.md
@@ -1,3 +1,7 @@
+# v1.229.1 (2026-01-09)
+
+* **Dependency Update**: Updated to the latest SDK module versions
+
 # v1.229.0 (2025-12-17)
 
 * **Feature**: Adding the newly launched p6-b300.48xlarge ec2  instance support in Sagemaker(Hyperpod,Training and Sceptor)