+++ service/vpclattice/CHANGELOG.md
@@ -1,3 +1,7 @@
+# v1.20.7 (2026-01-09)
+
+* **Dependency Update**: Updated to the latest SDK module versions
+
 # v1.20.6 (2025-12-08)
 
 * **Dependency Update**: Updated to the latest SDK module versions